More thorough gproc_dist:sync() + gproc_dist:get_leader()
gproc_dist:sync() now waits for the candidates to respond, ensuring that
when the call returns, any gproc update that happened before the call
will have been replicated to all participating nodes.

gproc_dist:get_leader() will ask the local gproc_dist server which node
is the leader. This is used in the gproc_dist_tests suite, but should be
useful in other contexts as well.