First implementation of environment variable handling.
The idea is to let gproc act as a cache for environment variables, and
at the same time provide a streamlined way of checking alternative sources
of configuration. Caching the values in gproc makes it easier to see what
each process assumes about its environment, and the option to inherit values
from a parent should make it easier to manage configuration.

First implementation. Tests still need to be written.