gproc:nb_wait/1 now works even if waiting on self.
Thanks to @idubrov for spotting. The notify_waiters function would delete all
reverse entries, including the own pid, even if this pid was the one registering.

My solution was to lift the reverse-entry removal from the notify_waiters fn,
and ensuring that it didn't remove the reverse entry of whatever pid actually
registered the name. Test cases added, both for local and distributed case.