Address issue where crashed process wasn't unregged on remote node
This was superficially due to a failed pattern-match in delete_globals(),
but the real issue was that the distr messages for orderly unreg and
cleanup after exit were different. A first quick attempt at fixing
this didn't fix everything, so a slightly laxer test in remove_entry/3
made the test suite pass. However, this means that there is still a
discrepancy that *might* be unsafe, and at the very least is ugly.